add z-index to captcha button
