+++ pages/[lng]/register/index.tsx
@@ -1,5 +1,5 @@
 /* library package */
-import { FC, useState } from 'react'
+import { FC, useState, useRef} from 'react'
 import { GetServerSideProps, InferGetServerSidePropsType } from 'next'
 import Router from 'next/router'
 import { toast } from 'react-toastify'
@@ -55,9 +55,16 @@ const RegisterPage: FC<any> = ({
   hasFacebookAuth
 }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
   const i18n: any = useI18n()
+  const recaptchaRef = useRef<any>()
 
   const [isVerified, setIsVerified] = useState<boolean>(false)
   const linksBreadcrumb = [`${i18n.t("header.home")}`, `${i18n.t("register.title")}`]
+
+  const getReCAPTCHAToken = async () => {
+    const token = await recaptchaRef.current.executeAsync()
+    recaptchaRef.current.reset()
+    return token
+  }
 
   return (
     <Layout
@@ -72,6 +79,7 @@ const RegisterPage: FC<any> = ({
         <div className={styles.register_container}>
           <LoginRegisterOTP
             type="register"
+            getReCAPTCHAToken={getReCAPTCHAToken}
             hasOtp={hasOtp}
             brand={brand}
             title={i18n.t("register.title")}
@@ -109,6 +117,11 @@ const RegisterPage: FC<any> = ({
           </LoginRegisterOTP>
         </div>
       </div>
+      <ReCAPTCHA
+        ref={recaptchaRef}
+        sitekey={process.env.NEXT_PUBLIC_SITEKEY_RECAPTCHA_INVISIBLE}
+        size='invisible'
+      />
     </Layout>
   )
 }
